update finetune
replace finetune.py with a branch in prep_run_dp_train, put  class modify_train_script into prep_run_dp_train, replace command "cp" with shutil.copy2
