+++ dpgen2/superop/__init__.py
@@ -1,9 +1,6 @@
 from .block import (
     ConcurrentLearningBlock,
 )
-from .finetune import (
-    Finetune,
-)
 from .prep_run_dp_train import (
     PrepRunDPTrain,
 )
