+++ dpgen2/superop/prep_run_dp_train.py
@@ -1,4 +1,5 @@
 import os
+import json
 from copy import (
     deepcopy,
 )
@@ -44,6 +45,80 @@ from dpgen2.utils.step_config import (
 )
 from dpgen2.utils.step_config import normalize as normalize_step_dict
 
+from dpgen2.constants import (
+    train_script_name,
+    train_task_pattern,
+)
+
+    
+class ModifyTrainScript(OP):
+    r"""Modify the training scripts to prepare them for training
+    tasks in dpgen step.
+
+    Read the training scripts modified by finetune, and replace
+    the original template scripts to be compatible with pre-trained models.
+    New templates are returned as `op["template_script"]`.
+
+    """
+
+    @classmethod
+    def get_input_sign(cls):
+        return OPIOSign(
+            {
+                "numb_models": int,
+                "scripts": Artifact(Path),
+            }
+        )
+
+    @classmethod
+    def get_output_sign(cls):
+        return OPIOSign(
+            {
+                "template_script": List[dict],
+            }
+        )
+
+    @OP.exec_sign_check
+    def execute(
+        self,
+        ip: OPIO,
+    ) -> OPIO:
+        r"""Execute the OP.
+
+        Parameters
+        ----------
+        ip : dict
+            Input dict with components:
+
+            - `scripts`: (`Artifact(Path)`) Training scripts from finetune.
+            - `numb_models`: (`int`) Number of DP models to train.
+
+        Returns
+        -------
+        op : dict
+            Output dict with components:
+
+            - `template_script`: (`List[dict]`) One template from one finetuning task. The length of the list should be the same as `numb_models`.
+
+        """
+        scripts = ip["scripts"]
+        new_template_script = []
+        numb_models = ip["numb_models"]
+
+        for ii in range(numb_models):
+            subdir = Path(train_task_pattern % ii)
+            train_script = Path(scripts) / subdir / train_script_name
+            with open(train_script, "r") as fp:
+                train_dict = json.load(fp)
+            new_template_script.append(train_dict)
+
+        op = OPIO(
+            {
+                "template_script": new_template_script,
+            }
+        )
+        return op
+
 
 class PrepRunDPTrain(Steps):
     def __init__(
@@ -54,6 +129,7 @@ class PrepRunDPTrain(Steps):
         prep_config: dict = normalize_step_dict({}),
         run_config: dict = normalize_step_dict({}),
         upload_python_packages: Optional[List[os.PathLike]] = None,
+        finetune: bool = False,
     ):
         self._input_parameters = {
             "block_id": InputParameter(type=str, value=""),
@@ -69,7 +145,9 @@ class PrepRunDPTrain(Steps):
             "init_data": InputArtifact(),
             "iter_data": InputArtifact(),
         }
-        self._output_parameters = {}
+        self._output_parameters = {
+            "template_script": OutputParameter(),
+        }
         self._output_artifacts = {
             "scripts": OutputArtifact(),
             "models": OutputArtifact(),
@@ -84,11 +162,14 @@ class PrepRunDPTrain(Steps):
                 artifacts=self._input_artifacts,
             ),
             outputs=Outputs(
+                parameters=self._output_parameters,
                 artifacts=self._output_artifacts,
             ),
         )
 
         self._keys = ["prep-train", "run-train"]
+        if finetune:
+            self._keys.append("modify-train-script")
         self.step_keys = {}
         ii = "prep-train"
         self.step_keys[ii] = "--".join(["%s" % self.inputs.parameters["block_id"], ii])
@@ -96,6 +177,8 @@ class PrepRunDPTrain(Steps):
         self.step_keys[ii] = "--".join(
             ["%s" % self.inputs.parameters["block_id"], ii + "-{{item}}"]
         )
+        ii = "modify-train-script"
+        self.step_keys[ii] = "--".join(["%s" % self.inputs.parameters["block_id"], ii])
 
         self = _prep_run_dp_train(
             self,
@@ -105,6 +188,7 @@ class PrepRunDPTrain(Steps):
             prep_config=prep_config,
             run_config=run_config,
             upload_python_packages=upload_python_packages,
+            finetune=finetune,
         )
 
     @property
@@ -136,6 +220,7 @@ def _prep_run_dp_train(
     prep_config: dict = normalize_step_dict({}),
     run_config: dict = normalize_step_dict({}),
     upload_python_packages: Optional[List[os.PathLike]] = None,
+    finetune: bool = False,
 ):
     prep_config = deepcopy(prep_config)
     run_config = deepcopy(run_config)
@@ -199,7 +284,33 @@ def _prep_run_dp_train(
         **run_config,
     )
     train_steps.add(run_train)
-
+    
+    if finetune is True:
+        modify_train_script = Step(
+            "modify-train-script",
+            template=PythonOPTemplate(
+                ModifyTrainScript,
+                python_packages=upload_python_packages,
+                **prep_template_config,
+            ),
+            parameters={
+                "numb_models": train_steps.inputs.parameters["numb_models"],
+            },
+            artifacts={
+                "scripts": run_train.outputs.artifacts["script"],
+            },
+            key=step_keys["modify-train-script"],
+            executor=prep_executor,
+            **prep_config,
+        )
+        train_steps.add(modify_train_script)
+        train_steps.outputs.parameters["template_script"].value_from_parameter = modify_train_script.outputs.parameters[
+            "template_script"
+        ]
+    else:
+        train_steps.outputs.parameters["template_script"].value_from_parameter = train_steps.inputs.parameters[
+            "template_script"
+        ]
     train_steps.outputs.artifacts["scripts"]._from = run_train.outputs.artifacts[
         "script"
     ]
